# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentación Geocerca - MMA API",
    "description": "Este servicio permite documentar los diferentes endpoints relacionados al MMA, especificamente utilizando las especificaciones swagger de OpenAPI 3.0.  \nSwagger : [https://swagger.io](https://swagger.io).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "tags": [
    {
      "name": "Coordenada",
      "description": "Muestra las coordenadas del lugar físico de un establecimiento."
    },

    {
      "name": "Establecimiento",
      "description": "Muestra los establecimientos pertenecientes al mma."
    },
    {
      "name": "Poligono",
      "description": "Muestra una geocerca referencial de los establecimientos."
    },
    {
      "name": "PuntoReferencia",
      "description": "Marca una posición especifica dentro de un predio del establecimiento."
    }
  ],
  "paths": {
    "/establecimientos": {
      "get": {
        "tags": ["Establecimiento"],
        "summary": "Este método muestra todos los establecimientos disponibles",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo_vu": {
                      "type": "integer",
                      "example": 1
                    },
                    "name_est": {
                      "type": "string",
                      "example": "Guacolda"
                    },
                    "latitud": {
                      "type": "long",
                      "example": "-23.06858082"
                    },
                    "longitud": {
                      "type": "long",
                      "example": "-70.3687702"
                    },
                    "este": {
                      "type": "long",
                      "example": "359788"
                    },
                    "norte": {
                      "type": "long",
                      "example": "7448232"
                    },
                    "huso": {
                      "type": "integer",
                      "example": "19"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener los establecimientos"
          }
        }
      }
    },

    "/establecimientos/{codigo_vu}": {
      "get": {
        "tags": ["Establecimiento"],
        "summary": "Este endpoints busca un establecimiento por codigo_vu",
        "operationId": "codigo_vu",
        "parameters": [
          {
            "name": "codigo_vu",
            "in": "path",
            "description": "codigo_vu del establecimiento que retornara",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo_vu": {
                      "type": "integer",
                      "example": 1
                    },
                    "name_est": {
                      "type": "string",
                      "example": "Guacolda"
                    },
                    "latitud": {
                      "type": "long",
                      "example": "-23.06858082"
                    },
                    "longitud": {
                      "type": "long",
                      "example": "-70.3687702"
                    },
                    "este": {
                      "type": "long",
                      "example": "359788"
                    },
                    "norte": {
                      "type": "long",
                      "example": "7448232"
                    },
                    "huso": {
                      "type": "integer",
                      "example": "19"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener el establecimiento"
          }
        }
      }
    },

    "/coordenadas/{codigo_vu}": {
      "get": {
        "tags": ["Coordenada"],
        "summary": "Este endpoints busca una coordenada por codigo_vu",
        "operationId": "codigo_vu",
        "parameters": [
          {
            "name": "codigo_vu",
            "in": "path",
            "description": "codigo_vu del establecimiento que retornara",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Coordenada"
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener el establecimiento"
          }
        }
      }
    },

    "/coordenadasBulk": {
      "post": {
        "tags": ["Coordenada"],
        "summary": "Cargar coordenadas",
        "requestBody": {
          "description": "Añadir Coordenadas",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/Coordenada"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Coordenada"
                }
              }
            }
          },
          "500": {
            "description": "Error al crear las coordenadas"
          }
        }
      }
    },

    "/poligonos": {
      "post": {
        "tags": ["Poligono"],
        "summary": "Cargar Poligono",
        "requestBody": {
          "description": "Añadir Poligonos",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/Poligono"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Poligono"
                }
              }
            }
          },
          "500": {
            "description": "Error al crear el poligono"
          }
        }
      }
    },

    "/allPoligonos": {
      "get": {
        "tags": ["Poligono"],
        "summary": "Este endpoints muestra todos los poligonos creados",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Poligono"
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener los poligonos"
          }
        }
      }
    },

    "/puntosReferencia/{codigo_vu}": {
      "get": {
        "tags": ["PuntoReferencia"],
        "summary": "Este endpoints busca un punto de referencia por medio de su codigo_vu.",
        "operationId": "codigo_vu",
        "parameters": [
          {
            "name": "codigo_vu",
            "in": "path",
            "description": "codigo_vu del punto que desea retornar",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/PuntoReferencia"
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener el establecimiento"
          }
        }
      }
    },

    "/puntosReferencia": {
      "post": {
        "tags": ["PuntoReferencia"],
        "summary": "Cargar Punto de referencia",
        "requestBody": {
          "description": "Añadir Puntos",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/PuntoReferencia"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/PuntoReferencia"
                }
              }
            }
          },
          "500": {
            "description": "Error al crear el punto de referencia"
          }
        }
      }
    }
  },

  "components": {
    "schemas": {
      "Establecimiento": {
        "type": "object",
        "properties": {
          "codigo_vu": {
            "type": "integer",
            "example": 6838
          },
          "name_est": {
            "type": "string",
            "example": "Guacolda"
          },
          "latitud": {
            "type": "long",
            "example": "-23.06858082"
          },
          "longitud": {
            "type": "long",
            "example": "-70.3687702"
          },
          "este": {
            "type": "long",
            "example": "359788"
          },
          "norte": {
            "type": "long",
            "example": "7448232"
          },
          "huso": {
            "type": "integer",
            "example": "19"
          }
        }
      },

      "Coordenada": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 2
          },
          "id_poligono": {
            "type": "integer",
            "example": 1
          },
          "codigo_vu": {
            "type": "integer",
            "example": 6638
          },
          "latitud": {
            "type": "long",
            "example": -28.462931831672577
          },

          "longitud": {
            "type": "long",
            "example": -71.2589674660051
          },
          "este": {
            "type": "long",
            "example": 278817.5518038366
          },
          "norte": {
            "type": "long",
            "example": 6849435.836324847
          }
        }
      },

      "Poligono": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 2
          },
          "codigo_vu": {
            "type": "integer",
            "example": 6838
          }
        }
      },

      "PuntoReferencia": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "codigo_vu": {
            "type": "integer",
            "example": 6638
          },
          "latitud": {
            "type": "long",
            "example": -28.462931831672577
          },

          "longitud": {
            "type": "long",
            "example": -71.2589674660051
          },
          "este": {
            "type": "long",
            "example": 278817.5518038366
          },
          "norte": {
            "type": "long",
            "example": 6849435.836324847
          }
        }
      }
    }
  }
}